pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "web-scan"
version = "1.0.0"
description = "TR-Pasif Web Güvenlik Skoru - Yasal ve etik pasif web güvenlik tarama aracı"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Koray Yolcu", email = "kkyolcu@gmail.com"}
]
keywords = ["security", "web-security", "passive-scan", "csp", "hsts", "turkish"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Internet :: WWW/HTTP",
]
dependencies = [
    "httpx>=0.26.0",
    "beautifulsoup4>=4.12.0",
    "rich>=13.7.0",
    "pydantic>=2.5.0",
    "dnspython>=2.4.0",
    "python-dateutil>=2.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "ruff>=0.1.0",
    "black>=23.12.0",
    "mypy>=1.7.0",
]


[project.scripts]
web-scan = "trscan.cli:main"

[project.urls]
Homepage = "https://github.com/koray-yolcu-sec/web-scan"
Documentation = "https://github.com/koray-yolcu-sec/web-scan#readme"
Repository = "https://github.com/koray-yolcu-sec/web-scan"
Issues = "https://github.com/koray-yolcu-sec/web-scan/issues"

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I", "N", "W", "UP"]

[tool.black]
line-length = 100
target-version = ['py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
